commit new files

+++ src/main/java/com/anton/eshop/data/OrderDetails.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -29,7 +30,7 @@ public class OrderDetails {
     @JoinColumn(name = "product_id")
     private Product product;
 
-    private int amountProducts;
+    private BigDecimal amountProducts;
 
     private double summa;
 }
